api-autogen/astria-api/pyproject.toml
[project]
name = "astria-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "ipython>=9.5.0",
    "neo4j>=5.28.2",
    "openapi-spec-validator>=0.7.2",
    "pyyaml>=6.0.2",
    "uvicorn>=0.35.0",
]
